Initial commit of ndeftool version 0.1.0
This is a Python 2 and Python 3 compatible rewrite of the ndeftool.py
example from https://github.com/nfcpy/nfcpy. It uses the ndeflib
module for decoding and encoding and the click module for implementing
the command line interface.

This version is regarded complete for core functionality (the load,
save, print, typename, identifier, and payload commands) and comes
with three commands for creating specific records (text, uri,
smartposter).